chore(slack, github, reddit): move request functions to single file
