--- github-repository.c
@@ -1,3 +1,0 @@
-/*
- * Code revolving around GitHub repositories.
- */
