+++ slack-socketmode.c
@@ -6,13 +6,12 @@
 #include "slack.h"
 #include "slack-internal.h"
 
-#include "cee-utils.h"
-
 static enum slack_sm_types
-eval_sm_type(char type[])
+sm_eval_type(char type[])
 {
   // EVENT API PAYLOADS
   if (STREQ("message", type)) return SLACK_SOCKETMODE_TYPE_MESSAGE;
+
   // INTERACTION PAYLOADS
   if (STREQ("block_actions", type)) return SLACK_SOCKETMODE_TYPE_BLOCK_ACTIONS;
   if (STREQ("message_actions", type))
@@ -20,106 +19,104 @@ eval_sm_type(char type[])
   if (STREQ("view_closed", type)) return SLACK_SOCKETMODE_TYPE_VIEW_CLOSED;
   if (STREQ("view_submission", type))
     return SLACK_SOCKETMODE_TYPE_VIEW_SUBMISSION;
+
   return SLACK_SOCKETMODE_TYPE_NONE;
 }
 
 static void
 send_acknowledge(struct slack_sm *sm, const char envelope_id[])
 {
-  char payload[512];
-  size_t ret =
-    json_inject(payload, sizeof(payload), "(envelope_id):s", envelope_id);
-  ASSERT_S(ret < sizeof(payload), "Out of bounds write attempt");
-
   struct ws_info info = { 0 };
-  ws_send_text(sm->ws, &info, payload, ret);
+  char payload[512];
+  size_t ret;
+
+  ret = json_inject(payload, sizeof(payload), "(envelope_id):s", envelope_id);
+  ASSERT_S(ret < sizeof(payload), "Out of bounds write attempt");
 
   log_info(
     ANSICOLOR("SEND",
               ANSI_FG_BRIGHT_GREEN) " ACKNOWLEDGE (%d bytes) [@@@_%zu_@@@]",
     ret, info.loginfo.counter);
+
+  ws_send_text(sm->ws, &info, payload, ret);
 }
 
 static void
 on_hello(struct slack_sm *sm, const char *text, size_t len)
 {
-  sm->is_ready          = true;
+  long interval_sec = 0;
+
+  sm->is_ready = true;
   sm->hbeat.interval_ms = 0;
-  sm->hbeat.tstamp      = cee_timestamp_ms();
+  sm->hbeat.tstamp = cee_timestamp_ms();
 
-  long interval_s = 0;
   json_extract((char *)text, len,
-               "(debug_info.approximate_connection_time):ld", &interval_s);
-  sm->hbeat.interval_ms = 1000 * interval_s;
+               "(debug_info.approximate_connection_time):ld", &interval_sec);
 
-  if (sm->cbs.on_hello) (*sm->cbs.on_hello)(sm->p_client, text, len);
+  sm->hbeat.interval_ms = 1000 * interval_sec;
+
+  if (sm->cbs.on_hello) sm->cbs.on_hello(CLIENT(sm, sm), text, len);
 }
 
 static void
 on_message(struct slack_sm *sm, struct sized_buffer *data)
 {
   if (sm->cbs.on_message)
-    (*sm->cbs.on_message)(sm->p_client, data->start, data->size);
+    sm->cbs.on_message(CLIENT(sm, sm), data->start, data->size);
 }
 
 static void
 on_block_actions(struct slack_sm *sm, struct sized_buffer *data)
 {
   if (sm->cbs.on_block_actions)
-    (*sm->cbs.on_block_actions)(sm->p_client, data->start, data->size);
+    sm->cbs.on_block_actions(CLIENT(sm, sm), data->start, data->size);
 }
 
 static void
 on_message_actions(struct slack_sm *sm, struct sized_buffer *data)
 {
   if (sm->cbs.on_message_actions)
-    (*sm->cbs.on_message_actions)(sm->p_client, data->start, data->size);
+    sm->cbs.on_message_actions(CLIENT(sm, sm), data->start, data->size);
 }
 
 static void
 on_view_closed(struct slack_sm *sm, struct sized_buffer *data)
 {
   if (sm->cbs.on_view_closed)
-    (*sm->cbs.on_view_closed)(sm->p_client, data->start, data->size);
+    sm->cbs.on_view_closed(CLIENT(sm, sm), data->start, data->size);
 }
 
 static void
 on_view_submission(struct slack_sm *sm, struct sized_buffer *data)
 {
   if (sm->cbs.on_view_submission)
-    (*sm->cbs.on_view_submission)(sm->p_client, data->start, data->size);
+    sm->cbs.on_view_submission(CLIENT(sm, sm), data->start, data->size);
 }
 
-static void *
+static void
 context_run(void *p_cxt)
 {
-  struct slack_event_cxt *cxt = p_cxt;
-  bool is_main_thread         = cxt->is_main_thread;
-  cxt->tid                    = pthread_self();
+  struct slack_event *cxt = p_cxt;
 
-  if (!is_main_thread)
-    log_info("Thread " ANSICOLOR("starts", ANSI_FG_RED) " to serve %s",
-             cxt->str_type);
+  log_info("Thread " ANSICOLOR("starts", ANSI_FG_RED) " to serve %s",
+           cxt->str_type);
 
-  if (cxt->on_event) (*cxt->on_event)(cxt->p_sm, &cxt->data);
+  cxt->on_event(cxt->sm, &cxt->data);
 
-  if (!is_main_thread) {
-    log_info("Thread " ANSICOLOR("exits", ANSI_FG_RED) " from serving %s",
-             cxt->str_type);
+  log_info("Thread " ANSICOLOR("exits", ANSI_FG_RED) " from serving %s",
+           cxt->str_type);
 
-    free(cxt->data.start);
-    free(cxt);
-    pthread_exit(NULL);
-  }
-  return NULL;
+  free(cxt->data.start);
+  free(cxt);
 }
 
 static void
 on_events(struct slack_sm *sm, struct sized_buffer *data, char str_type[])
 {
   void (*on_event)(struct slack_sm *, struct sized_buffer *) = NULL;
-  enum slack_sm_types type = eval_sm_type(str_type);
-  switch (type) {
+  enum slack_sm_types type;
+
+  switch (type = sm_eval_type(str_type)) {
   case SLACK_SOCKETMODE_TYPE_MESSAGE:
     if (sm->cbs.on_message) on_event = &on_message;
     break;
@@ -140,37 +137,33 @@ on_events(struct slack_sm *sm, struct sized_buffer *data, char str_type[])
     break;
   }
 
-  if (!on_event) return;
+  if (on_event) {
+    enum slack_event_handling_mode mode;
 
-  struct slack_event_cxt cxt;
-  asprintf(&cxt.data.start, "%.*s", (int)data->size, data->start);
-  cxt.data.size = data->size;
-  cxt.p_sm      = sm;
-  cxt.type      = type;
-  cxt.on_event  = on_event;
-  snprintf(cxt.str_type, sizeof(cxt.str_type), "%s", str_type);
+    switch (mode = sm->event_handler(CLIENT(sm, sm), data, type)) {
+    case SLACK_EVENT_IGNORE:
+      return;
+    case SLACK_EVENT_MAIN_THREAD:
+      on_event(sm, data);
+      return;
+    case SLACK_EVENT_CHILD_THREAD: {
+      struct slack_event *cxt = malloc(sizeof *cxt);
+      int ret;
 
-  enum slack_event_handling_mode mode =
-    sm->event_handler(sm->p_client, &cxt.data, cxt.type);
-  switch (mode) {
-  case SLACK_EVENT_IGNORE:
-    return;
-  case SLACK_EVENT_MAIN_THREAD:
-    cxt.is_main_thread = true;
-    context_run(&cxt);
-    return;
-  case SLACK_EVENT_CHILD_THREAD: {
-    cxt.is_main_thread            = false;
-    struct slack_event_cxt *p_cxt = malloc(sizeof *p_cxt);
-    memcpy(p_cxt, &cxt, sizeof(cxt));
-    pthread_t tid;
-    if (pthread_create(&tid, NULL, &context_run, p_cxt))
-      ERR("Couldn't create thread");
-    if (pthread_detach(tid)) ERR("Couldn't detach thread");
-    return;
-  }
-  default:
-    ERR("Unknown event handling mode (code: %d)", mode);
+      cxt->data.size =
+        asprintf(&cxt->data.start, "%.*s", (int)data->size, data->start);
+      cxt->sm = sm;
+      cxt->type = type;
+      cxt->on_event = on_event;
+      snprintf(cxt->str_type, sizeof(cxt->str_type), "%s", str_type);
+
+      ret = work_run(&context_run, cxt);
+      VASSERT_S(0 == ret, "Couldn't create task (code %d)", ret);
+    }
+      return;
+    default:
+      ERR("Unknown event handling mode (code: %d)", mode);
+    }
   }
 }
 
@@ -208,7 +201,11 @@ on_text_cb(void *p_sm,
 {
   struct slack_sm *sm = p_sm;
 
+  struct sized_buffer data = { 0 };
+  char event_type[64] = "";
+  bool accepts_response_payload = false;
   char type[64] = "", envelope_id[64] = "";
+
   json_extract((char *)text, len, "(type):s,(envelope_id):s", type,
                envelope_id);
 
@@ -219,9 +216,6 @@ on_text_cb(void *p_sm,
     return;
   }
 
-  struct sized_buffer data      = { 0 };
-  char event_type[64]           = "";
-  bool accepts_response_payload = false;
   if (STREQ(type, "events_api")) {
     json_extract((char *)text, len,
                  "(payload):T"
@@ -256,13 +250,15 @@ noop_event_handler(struct slack *a,
 static void
 refresh_connection(struct slack_sm *sm)
 {
-  struct sized_buffer resp_body = { 0 };
-  slack_apps_connections_open(sm->p_client, &resp_body);
 
-  bool status                  = false;
-  char *base_url               = NULL;
+  bool status = false;
+  char *base_url = NULL;
+  struct sized_buffer body = { 0 };
   struct sized_buffer messages = { 0 };
-  json_extract(resp_body.start, resp_body.size,
+
+  slack_apps_connections_open(CLIENT(sm, sm), &body);
+
+  json_extract(body.start, body.size,
                "(ok):b, (url):?s, (response_metadata.messages):T", &status,
                &base_url, &messages);
 
@@ -270,23 +266,23 @@ refresh_connection(struct slack_sm *sm)
             "Couldn't fetch connections for websockets:\n\t\tMessage: %.*s",
             (int)messages.size, messages.start);
 
-  ws_set_url(sm->p_client->sm.ws, base_url, NULL);
+  ws_set_url(sm->ws, base_url, NULL);
 
   free(base_url);
-  free(resp_body.start);
+  free(body.start);
 }
 
 void
 slack_sm_init(struct slack_sm *sm, struct logconf *conf)
 {
-  struct ws_callbacks cbs = { .data       = sm,
+  struct ws_callbacks cbs = { .data = sm,
                               .on_connect = &on_connect_cb,
-                              .on_text    = &on_text_cb,
-                              .on_close   = &on_close_cb };
-  struct ws_attr attr     = { .conf = conf };
+                              .on_text = &on_text_cb,
+                              .on_close = &on_close_cb };
+  struct ws_attr attr = { .conf = conf };
 
   sm->mhandle = curl_multi_init();
-  sm->ws      = ws_init(&cbs, sm->mhandle, &attr);
+  sm->ws = ws_init(&cbs, sm->mhandle, &attr);
   logconf_branch(&sm->conf, conf, "SLACK_SOCKETMODE");
 
   sm->event_handler = &noop_event_handler;
