+++ slack-webapi.c
@@ -6,16 +6,32 @@
 
 #include "slack.h"
 #include "slack-internal.h"
-#include "cee-utils.h"
 
-#define SLACK_BASE_API_URL "https://slack.com/api"
+static void
+setopt_cb(struct ua_conn *conn, void *p_token)
+{
+  struct sized_buffer *token = p_token;
+  char auth[128];
+  int ret;
+
+  ret = snprintf(auth, sizeof(auth), "Bearer %.*s", (int)token->size,
+                 token->start);
+  ASSERT_S(ret < sizeof(auth), "Out of bounds write attempt");
+
+  ua_conn_add_header(conn, "Authorization", auth);
+  ua_conn_add_header(webapi->ua, "Content-type",
+                     "application/x-www-form-urlencoded");
+
+#if 0 /* enable for debugging */
+  curl_easy_setopt(ua_conn_get_easy_handle(conn), CURLOPT_VERBOSE, 1L);
+#endif
+}
 
 void
 slack_webapi_init(struct slack_webapi *webapi,
                   struct logconf *conf,
                   struct sized_buffer *token)
 {
-#if 0
   webapi->ua = ua_init(&(struct ua_attr){ .conf = conf });
   ua_set_url(webapi->ua, SLACK_BASE_API_URL);
   logconf_branch(&webapi->conf, conf, "SLACK_WEBAPI");
@@ -25,15 +41,7 @@ slack_webapi_init(struct slack_webapi *webapi,
   }
   ASSERT_S(NULL != token->start, "Missing bot token");
 
-  char auth[128];
-  int ret = snprintf(auth, sizeof(auth), "Bearer %.*s", (int)token->size,
-                     token->start);
-  ASSERT_S(ret < sizeof(auth), "Out of bounds write attempt");
-
-  ua_reqheader_add(webapi->ua, "Authorization", auth);
-  ua_reqheader_add(webapi->ua, "Content-type",
-                   "application/x-www-form-urlencoded");
-#endif
+  ua_set_opt(adapter->ua, token, &setopt_cb);
 }
 
 void
@@ -45,21 +53,18 @@ slack_webapi_cleanup(struct slack_webapi *webapi)
 /* template function for performing requests */
 ORCAcode
 slack_webapi_run(struct slack_webapi *webapi,
-                 struct sized_buffer *ret,
+                 struct slack_request_attr *attr,
                  struct sized_buffer *body,
                  enum http_method method,
                  char endpoint_fmt[],
                  ...)
 {
-  struct ua_resp_handle handle = { ret
-                                     ? (void (*)(char *, size_t, void *))
-                                         & cee_sized_buffer_from_json
-                                     : NULL,
-                                   ret };
-  struct ua_conn_attr conn_attr = { 0 };
   char endpoint[2048];
   va_list args;
   size_t len;
+
+  struct ua_resp_handle handle = { attr->obj, 0, NULL, attr->from_json };
+  struct ua_conn_attr conn_attr = { method, body, endpoint };
 
   va_start(args, endpoint_fmt);
 
@@ -68,9 +73,117 @@ slack_webapi_run(struct slack_webapi *webapi,
 
   va_end(args);
 
-  conn_attr.method = method;
-  conn_attr.body = body;
-  conn_attr.endpoint = endpoint;
-
   return ua_easy_run(webapi->ua, NULL, &handle, &conn_attr);
+}
+
+ORCAcode
+slack_apps_connections_open(struct slack *client, struct sized_buffer *ret)
+{
+  ORCA_EXPECT(client, !IS_EMPTY_STRING(client->bot_token.start), ORCA_BAD_PARAMETER);
+  ORCA_EXPECT(client, !IS_EMPTY_STRING(client->app_token.start), ORCA_BAD_PARAMETER);
+
+  char auth[128] = "";
+  size_t len;
+
+  len = snprintf(auth, sizeof(auth), "Bearer %.*s",
+                 (int)client->app_token.size, client->app_token.start);
+  ASSERT_S(len < sizeof(auth), "Out of bounds write attempt");
+  ua_reqheader_add(client->webapi.ua, "Authorization", auth);
+
+  ORCAcode code;
+  code = slack_webapi_run(&client->webapi, ret, NULL, HTTP_POST,
+                          "/apps.connections.open");
+
+  len = snprintf(auth, sizeof(auth), "Bearer %.*s",
+                 (int)client->bot_token.size, client->bot_token.start);
+  ASSERT_S(len < sizeof(auth), "Out of bounds write attempt");
+  ua_reqheader_add(client->webapi.ua, "Authorization", auth);
+
+  return code;
+}
+
+ORCAcode
+slack_auth_test(struct slack *client, struct sized_buffer *ret)
+{
+  return slack_webapi_run(&client->webapi, ret, NULL, HTTP_POST,
+                          "/auth.test");
+}
+
+ORCAcode
+slack_chat_post_message(struct slack *client,
+                        struct slack_chat_post_message_params *params,
+                        struct sized_buffer *ret)
+{
+#if 0
+  ORCA_EXPECT(client, params != NULL, ORCA_BAD_PARAMETER);
+  ORCA_EXPECT(client, !IS_EMPTY_STRING(params->channel), ORCA_BAD_PARAMETER);
+
+  char *payload = NULL;
+  size_t len = json_ainject(&payload,
+                            "(token):s"
+                            "(channel):s"
+#if 0
+                "(as_user):b"
+#endif
+                            "(icon_url):s"
+                            "(icon_emoji):s"
+                            "(text):s"
+                            "(thread_ts):s"
+                            "(username):s",
+                            params->token, params->channel,
+#if 0
+                &params->as_user,
+#endif
+                            params->icon_url, params->icon_emoji, params->text,
+                            params->thread_ts, params->username);
+
+  if (!payload) {
+    log_error("Couldn't create payload");
+    return ORCA_BAD_PARAMETER;
+  }
+
+  ua_reqheader_add(client->webapi.ua, "Content-type", "application/json");
+
+  ORCAcode code;
+  code = slack_webapi_run(&client->webapi, ret,
+                          &(struct sized_buffer){ payload, len }, HTTP_POST,
+                          "/chat.postMessage");
+
+  ua_reqheader_add(client->webapi.ua, "Content-type",
+                   "application/x-www-form-urlencoded");
+
+  free(payload);
+
+  return code;
+#else
+  return -1;
+#endif
+}
+
+ORCAcode
+slack_users_info(struct slack *client,
+                 struct slack_users_info_params *params,
+                 struct sized_buffer *ret)
+{
+  ORCA_EXPECT(client, params != NULL, ORCA_BAD_PARAMETER);
+  ORCA_EXPECT(client, !IS_EMPTY_STRING(params->user), ORCA_BAD_PARAMETER);
+
+  char query[4096];
+  size_t len = 0;
+
+  len += snprintf(query + len, sizeof(query) - len, "user=%s", params->user);
+  ASSERT_S(len < sizeof(query), "Out of bounds write attempt");
+  if (params->token) {
+    len +=
+      snprintf(query + len, sizeof(query) - len, "&token=%s", params->token);
+    ASSERT_S(len < sizeof(query), "Out of bounds write attempt");
+  }
+  if (params->include_locale) {
+    len += snprintf(query + len, sizeof(query) - len, "&include_locale=true");
+    ASSERT_S(len < sizeof(query), "Out of bounds write attempt");
+  }
+
+  return slack_webapi_run(&client->webapi, ret,
+                          &(struct sized_buffer){ query, len }, HTTP_POST,
+                          "/users.info");
 }
